rust: avoid packed structs.
This leads to very bad code on RISC-V, because the compiler seems to
assume that packed structs are unaligned and thus performs all memory
accesses byte-wise (!!). I think this is also a Rust problem, because
it does not allow to make structs both aligned and packed.

+++ src/include/base/PEDesc.h
@@ -61,7 +61,7 @@ static constexpr size_t isa_count() {
  * Describes a PE
  */
 struct PEDesc {
-    typedef uint32_t value_t;
+    typedef uint64_t value_t;
 
     /**
      * Default constructor
