+++ src/include/base/arch/gem5/Env.h
@@ -38,8 +38,8 @@ public:
 
 struct BootEnv {
     uint64_t pe_id;
-    uint32_t pe_desc;
-    uint32_t argc;
+    uint64_t pe_desc;
+    uint64_t argc;
     uint64_t argv;
     uint64_t heap_size;
     uint64_t pe_mem_base;
